+++ Assets/Room02_Operating/Ending/EscapeChaseQTE.cs
@@ -859,7 +859,7 @@ namespace EscapeRoom
             tmp.fontSize = fontSize;
             tmp.alignment = TextAlignmentOptions.Center;
             tmp.color = Color.white;
-            tmp.enableWordWrapping = true;
+            tmp.textWrappingMode = TextWrappingModes.Normal;
             return tmp;
         }
     }